-. add header for balance

+++ entity/bill.go
@@ -10,14 +10,14 @@ type AddBillReq struct {
 	Items           []ItemPurchase `json:"bill_items"`
 	BankName        string         `json:"bill_bank_name"`
 	AccountNumber   string         `json:"bill_account_number"`
-	ShippingCost    int32          `json:"bill_shipping_cost"`
+	ShippingCost    float64        `json:"bill_shipping_cost"`
 	BillNote        *string        `json:"bill_notes"`
 }
 
 type ItemPurchase struct {
-	ItemId       int32  `json:"item_id"`
-	ItemQty      int32  `json:"item_qty"`
-	ItemDiscount *int32 `json:"item_discount"`
+	ItemId       int32    `json:"item_id"`
+	ItemQty      int32    `json:"item_qty"`
+	ItemDiscount *float64 `json:"item_discount"`
 }
 
 type BillDetailsResp struct {
@@ -31,21 +31,21 @@ type BillDetailsResp struct {
 	BillStatus       string       `json:"bill_status"`
 	BillSubTotal     int64        `json:"bill_subtotal"`
 	BillTotal        int64        `json:"bill_total"`
-	BillShippingCost int32        `json:"bill_shipping_cost"`
+	BillShippingCost float64      `json:"bill_shipping_cost"`
 }
 
 type ItemBill struct {
 	Name        string  `json:"item_name"`
 	Description *string `json:"item_description"`
 	Qty         int32   `json:"item_qty"`
-	Price       int32   `json:"item_price"`
-	Amount      int32   `json:"item_amount"`
+	Price       float64 `json:"item_price"`
+	Amount      float64 `json:"item_amount"`
 }
 
 type BillHeaderResp struct {
-	Overdue   int32 `json:"bill_overdue"`
-	Open      int32 `json:"bill_open"`
-	BillDraft int32 `json:"bill_draft"`
+	Overdue   float64 `json:"bill_overdue"`
+	Open      float64 `json:"bill_open"`
+	BillDraft float64 `json:"bill_draft"`
 }
 
 type BillUpdateStatusReq struct {
