+++ model/bills.gen.go
@@ -18,10 +18,10 @@ type Bill struct {
 	BillDueDate       time.Time `gorm:"column:bill_due_date;not null" json:"bill_due_date"`
 	BillNumber        string    `gorm:"column:bill_number;not null" json:"bill_number"`
 	BillOrderNumber   *string   `gorm:"column:bill_order_number" json:"bill_order_number"`
-	BillTotal         int32     `gorm:"column:bill_total;not null" json:"bill_total"`
+	BillTotal         float64   `gorm:"column:bill_total;not null" json:"bill_total"`
 	BillStatus        string    `gorm:"column:bill_status;not null" json:"bill_status"`
 	BillType          string    `gorm:"column:bill_type;not null" json:"bill_type"`
-	BillShippingCost  int32     `gorm:"column:bill_shipping_cost;not null" json:"bill_shipping_cost"`
+	BillShippingCost  float64   `gorm:"column:bill_shipping_cost;not null" json:"bill_shipping_cost"`
 	BillAccountNumber string    `gorm:"column:bill_account_number;not null" json:"bill_account_number"`
 	BillBankName      string    `gorm:"column:bill_bank_name;not null" json:"bill_bank_name"`
 	BillNotes         *string   `gorm:"column:bill_notes" json:"bill_notes"`
