+++ model/items.gen.go
@@ -12,8 +12,8 @@ type Item struct {
 	SupplierID        int32   `gorm:"column:supplier_id;not null" json:"supplier_id"`
 	ItemName          string  `gorm:"column:item_name;not null" json:"item_name"`
 	ItemDescription   *string `gorm:"column:item_description" json:"item_description"`
-	ItemPurchasePrice int32   `gorm:"column:item_purchase_price;not null" json:"item_purchase_price"`
-	ItemSellPrice     int32   `gorm:"column:item_sell_price;not null" json:"item_sell_price"`
+	ItemPurchasePrice float64 `gorm:"column:item_purchase_price;not null" json:"item_purchase_price"`
+	ItemSellPrice     float64 `gorm:"column:item_sell_price;not null" json:"item_sell_price"`
 	ItemUnit          string  `gorm:"column:item_unit;not null" json:"item_unit"`
 }
